Load Textures using the Assets class
Assets holds RefPtrs of Textures, mainly because I can't create an
`ErrorOr<Texture&>`. This means that once a Texture is loaded, it will
remain loaded forever. For now, that's good enough. Eventually it would
be good to be able to evict assets that aren't in use (and haven't been
used in a while) when we run low on memory, but that's extra complexity
we don't need right now. In fact we could always do that now by just
seeing if a Texture has a ref-count of 1. So yeah it's fine.